fix launch methods working; add import coroutines.collect; create assert directory, enum class Command is moved outside the view model;

+++ app/src/main/java/com/geras/punk_rockplayer/data_source/Song.kt
@@ -5,5 +5,5 @@ data class Song(
     val artist: String = "",
     val bitmapUri: String = "",
     val trackUri: String = "",
-    val duration: Int = 0
+    val duration: Long = 0
 )
